--- .eslintrc.yml
@@ -1,7 +1,0 @@
-env:
-  jest: true
-  node: true
-parser: babel-eslint
-extends:
-  - eslint:recommended
-  - plugin:prettier/recommended
