+++ package.json
@@ -10,11 +10,12 @@
     "node": "=16.x"
   },
   "scripts": {
-    "lint": "eslint .",
+    "lint": "eslint . --ext .ts",
+    "lint-fix": "npm run lint -- --fix",
     "test": "mocha --require ts-node/register --extensions ts tests 'tests/**/*.ts' --exit",
     "test-inspect": "mocha --inspect-brk --require ts-node/register --extensions ts tests 'tests/**/*.ts' --exit",
     "test-watch": "mocha --require ts-node/register --watch --watch-files src --watch-files tests --extensions ts tests 'tests/**/*.ts'",
-    "start": "node -r ts-node/register trial.ts"
+    "start": "node -r ts-node/register src/trial.ts"
   },
   "devDependencies": {
     "@types/chai": "^4.3.0",
@@ -24,9 +25,10 @@
     "@typescript-eslint/parser": "^5.13.0",
     "chai": "^4.3.6",
     "eslint": "^8.10.0",
+    "eslint-config-airbnb": "^19.0.4",
+    "eslint-config-airbnb-typescript": "^16.1.1",
     "mocha": "^9.2.1",
     "nodemon": "^2.0.15",
-    "prettier": "^1.14.3",
     "ts-node": "^10.6.0",
     "typescript": "^4.6.2"
   },
