+++ src/domain/models/Drug.ts
@@ -1,11 +1,16 @@
-import { differenceInDays } from "date-fns";
+import { differenceInDays } from 'date-fns';
 
 export class Drug {
   protected readonly name: string;
+
   protected readonly originalExpiresIn: number; // number of days
+
   protected readonly originalBenefit: number; // how powerful the drug is
+
   protected readonly timeBenefitFactor: number;
+
   protected readonly timeBenefitFactorAfterExpiration: number;
+
   protected readonly createdAt: Date;
 
   constructor(params: {
@@ -60,14 +65,12 @@ export class Drug {
 
     const {
       numberOfDaysAfterExpiration,
-      numberOfDaysBeforeExpiration
+      numberOfDaysBeforeExpiration,
     } = this.getBenefitAtCalculationParams(targetDate);
 
-    const benefitAt = this.originalBenefit +
-      (numberOfDaysBeforeExpiration * this.timeBenefitFactor) +
-      numberOfDaysAfterExpiration * this.timeBenefitFactorAfterExpiration
-    ;
-
+    const benefitAt = this.originalBenefit
+      + (numberOfDaysBeforeExpiration * this.timeBenefitFactor)
+      + numberOfDaysAfterExpiration * this.timeBenefitFactorAfterExpiration;
     return this.benefitValueGuard(benefitAt);
   }
 
@@ -80,7 +83,7 @@ export class Drug {
 
     let numberOfDaysAfterExpiration = diffInDays - this.originalExpiresIn || 0;
     if (numberOfDaysAfterExpiration < 0) {
-      numberOfDaysAfterExpiration = 0
+      numberOfDaysAfterExpiration = 0;
     }
 
     const numberOfDaysBeforeExpiration = diffInDays - numberOfDaysAfterExpiration;
@@ -88,11 +91,11 @@ export class Drug {
     return {
       diffInDays,
       numberOfDaysAfterExpiration,
-      numberOfDaysBeforeExpiration
-    }
+      numberOfDaysBeforeExpiration,
+    };
   }
 
-  protected benefitValueGuard (benefit: number): number {
+  protected benefitValueGuard(benefit: number): number {
     if (benefit < 0) {
       return 0;
     }
@@ -122,68 +125,5 @@ export class Drug {
     if (!(params.createdAt instanceof Date)) {
       throw new Error(`createdAt "${params.createdAt}" must be a Date`);
     }
-  }
-}
-
-export class Fervex extends Drug {
-  constructor(params: {
-    originalExpiresIn: number;
-    originalBenefit: number;
-    createdAt: Date;
-  }) {
-    super({
-      name: 'Fervex',
-      originalExpiresIn: params.originalExpiresIn,
-      originalBenefit: params.originalBenefit,
-      createdAt: params.createdAt,
-    }, { timeBenefitFactor: 1 });
-  }
-
-  /**
-   * override Drug for specific behavior
-   * @param targetDate
-   */
-  getBenefitAt(targetDate: Date): number {
-    if (targetDate < this.getCreatedAt()) {
-      return this.originalBenefit; // not alterated in the past. May be discussable.
-    }
-
-    const { diffInDays } = this.getBenefitAtCalculationParams(targetDate);
-
-    const diffBeforeExpiration = this.getExpiresIn() - diffInDays;
-
-    if (diffInDays > this.getExpiresIn()) {
-      return 0;
-    }
-
-    if (diffBeforeExpiration > 10) {
-      return super.getBenefitAt(targetDate);
-    }
-
-    if (this.originalExpiresIn <= 5) {
-      return this.benefitValueGuard(this.originalBenefit + diffInDays * 3);
-    }
-
-    const lastThreshold = this.originalExpiresIn - 5;
-    let numberOfDaysLastThreshold = diffInDays - lastThreshold;
-    if (numberOfDaysLastThreshold <= 0) {
-      numberOfDaysLastThreshold = 0;
-    }
-
-    const secondThreshold = this.originalExpiresIn - 10;
-    let numberOfDaysSecondThreshold = diffInDays - secondThreshold - numberOfDaysLastThreshold;
-    if (numberOfDaysSecondThreshold <= 0) {
-      numberOfDaysSecondThreshold = 0;
-    }
-
-    const numberOfDaysFirstThreshold = diffInDays - numberOfDaysSecondThreshold - numberOfDaysLastThreshold;
-
-    const benefit = this.originalBenefit +
-      (numberOfDaysLastThreshold * 3) +
-      numberOfDaysSecondThreshold * 2 +
-      numberOfDaysFirstThreshold
-    ;
-
-    return this.benefitValueGuard(benefit);
   }
 }
